SUCESSO KRLLLLLLL

+++ Packet.h
@@ -109,7 +109,7 @@ char *get_type_packet(unsigned char* buffer);
 /* ================ funcoes auxiliares ================ */
 void read_packet(unsigned char *buffer);
 int is_our_packet(unsigned char *buffer);
-int is_our_mask(unsigned long *buffer);
+int is_our_mask(unsigned short *buffer);
 int is_valid_type(int tipo);
 void print_bytes(char *nome,unsigned char *buf, int n);
